Add files via upload
This  are having the information of the student. in that data some missing values are there. in this file we can move to the database that cant support this file . due to file having the missing values. by using the ml algaritham   we have to minimize the errore values. that purpuse we need this file.